# contracts/taskbox.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TaskBox API",
    "version": "1.0.0",
    "description": "A project task API used throughout the API Engineering course."
  },
  "servers": [{"url": "http://localhost:8000", "description": "Local development"}],
  "tags": [
    {"name": "Health"},
    {"name": "Auth"},
    {"name": "Projects"},
    {"name": "Tasks"},
    {"name": "Webhooks"}
  ],
  "paths": {
    "/healthz": {
      "get": {"operationId": "healthCheck", "tags": ["Health"], "responses": {"200": {"description": "Service is healthy", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Health"}}}}}}
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "registerUser",
        "tags": ["Auth"],
        "summary": "Create a user account",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegisterRequest"}}}},
        "responses": {"201": {"description": "User created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "409": {"$ref": "#/components/responses/Conflict"}, "422": {"$ref": "#/components/responses/ValidationError"}}
      }
    },
    "/api/v1/auth/token": {
      "post": {
        "operationId": "createToken",
        "tags": ["Auth"],
        "summary": "Exchange credentials for a JWT",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}},
        "responses": {"200": {"description": "Access token", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenResponse"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "422": {"$ref": "#/components/responses/ValidationError"}}
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "getCurrentUser",
        "tags": ["Auth"],
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Current user", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}}
      }
    },
    "/api/v1/projects": {
      "get": {
        "operationId": "listProjects",
        "tags": ["Projects"],
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/Cursor"}, {"$ref": "#/components/parameters/Limit"}],
        "responses": {"200": {"description": "Projects visible to the current user", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectPage"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}}
      },
      "post": {
        "operationId": "createProject",
        "tags": ["Projects"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectCreate"}}}},
        "responses": {"201": {"description": "Project created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Project"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "422": {"$ref": "#/components/responses/ValidationError"}}
      }
    },
    "/api/v1/projects/{project_id}": {
      "parameters": [{"$ref": "#/components/parameters/ProjectId"}],
      "get": {
        "operationId": "getProject",
        "tags": ["Projects"],
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Project", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Project"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "404": {"$ref": "#/components/responses/NotFound"}}
      },
      "patch": {
        "operationId": "updateProject",
        "tags": ["Projects"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProjectUpdate"}}}},
        "responses": {"200": {"description": "Updated project", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Project"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/ValidationError"}}
      },
      "delete": {
        "operationId": "deleteProject",
        "tags": ["Projects"],
        "security": [{"bearerAuth": []}],
        "responses": {"204": {"description": "Project deleted"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}}
      }
    },
    "/api/v1/projects/{project_id}/members": {
      "parameters": [{"$ref": "#/components/parameters/ProjectId"}],
      "get": {
        "operationId": "listMembers",
        "tags": ["Projects"],
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/Cursor"}, {"$ref": "#/components/parameters/Limit"}],
        "responses": {"200": {"description": "Project members", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MembershipPage"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}}
      },
      "post": {
        "operationId": "addMember",
        "tags": ["Projects"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MembershipCreate"}}}},
        "responses": {"201": {"description": "Member added", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Membership"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}, "409": {"$ref": "#/components/responses/Conflict"}}
      }
    },
    "/api/v1/projects/{project_id}/members/{user_id}": {
      "parameters": [{"$ref": "#/components/parameters/ProjectId"}, {"$ref": "#/components/parameters/UserId"}],
      "patch": {
        "operationId": "updateMemberRole",
        "tags": ["Projects"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MembershipUpdate"}}}},
        "responses": {"200": {"description": "Updated membership", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Membership"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}}
      },
      "delete": {
        "operationId": "removeMember",
        "tags": ["Projects"],
        "security": [{"bearerAuth": []}],
        "responses": {"204": {"description": "Member removed"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}}
      }
    },
    "/api/v1/projects/{project_id}/tasks": {
      "parameters": [{"$ref": "#/components/parameters/ProjectId"}],
      "get": {
        "operationId": "listTasks",
        "tags": ["Tasks"],
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/Cursor"}, {"$ref": "#/components/parameters/Limit"}, {"$ref": "#/components/parameters/TaskStatus"}, {"name": "assignee_id", "in": "query", "schema": {"type": "string", "format": "uuid"}}],
        "responses": {"200": {"description": "Tasks ordered by created_at and id", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskPage"}}}}, "400": {"$ref": "#/components/responses/InvalidCursor"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}}
      },
      "post": {
        "operationId": "createTask",
        "tags": ["Tasks"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskCreate"}}}},
        "responses": {"201": {"description": "Task created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/ValidationError"}}
      }
    },
    "/api/v1/tasks/{task_id}": {
      "parameters": [{"$ref": "#/components/parameters/TaskId"}],
      "get": {
        "operationId": "getTask",
        "tags": ["Tasks"],
        "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Task", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}}
      },
      "patch": {
        "operationId": "updateTask",
        "tags": ["Tasks"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskUpdate"}}}},
        "responses": {"200": {"description": "Updated task", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Task"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}, "422": {"$ref": "#/components/responses/ValidationError"}}
      },
      "delete": {
        "operationId": "deleteTask",
        "tags": ["Tasks"],
        "security": [{"bearerAuth": []}],
        "responses": {"204": {"description": "Task deleted"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"$ref": "#/components/responses/Forbidden"}, "404": {"$ref": "#/components/responses/NotFound"}}
      }
    },
    "/api/v1/webhooks/tasks/import": {
      "post": {
        "operationId": "importTasksWebhook",
        "tags": ["Webhooks"],
        "summary": "Import tasks from a signed, idempotent webhook",
        "parameters": [{"name": "X-Webhook-Event-Id", "in": "header", "required": true, "schema": {"type": "string"}}, {"name": "X-Webhook-Signature", "in": "header", "required": true, "schema": {"type": "string", "description": "HMAC-SHA256 signature of the raw request body"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WebhookImportRequest"}}}},
        "responses": {"202": {"description": "Tasks accepted", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WebhookImportResponse"}}}}, "400": {"$ref": "#/components/responses/ValidationError"}, "401": {"$ref": "#/components/responses/InvalidSignature"}, "409": {"$ref": "#/components/responses/Conflict"}}
      }
    }
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}},
    "parameters": {
      "ProjectId": {"name": "project_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "TaskId": {"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "UserId": {"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
      "Cursor": {"name": "cursor", "in": "query", "required": false, "schema": {"type": "string"}, "description": "Opaque cursor returned as next_cursor."},
      "Limit": {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 50}},
      "TaskStatus": {"name": "status", "in": "query", "required": false, "schema": {"$ref": "#/components/schemas/TaskStatus"}}
    },
    "responses": {
      "Unauthorized": {"description": "Authentication is required", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetail"}}}},
      "Forbidden": {"description": "The authenticated user lacks the required project role", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetail"}}}},
      "NotFound": {"description": "Resource not found", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetail"}}}},
      "Conflict": {"description": "Resource conflicts with existing state", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetail"}}}},
      "ValidationError": {"description": "Request validation failed", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetail"}}}},
      "InvalidCursor": {"description": "Cursor is invalid or expired", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetail"}}}},
      "InvalidSignature": {"description": "Webhook signature is invalid", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/ProblemDetail"}}}}
    },
    "schemas": {
      "Health": {"type": "object", "required": ["status"], "properties": {"status": {"type": "string", "const": "ok"}}},
      "ProblemDetail": {"type": "object", "required": ["type", "title", "status", "detail"], "properties": {"type": {"type": "string", "format": "uri-reference"}, "title": {"type": "string"}, "status": {"type": "integer"}, "detail": {"type": "string"}, "instance": {"type": "string"}, "code": {"type": "string"}, "errors": {"type": "array", "items": {"type": "object"}}}},
      "RegisterRequest": {"type": "object", "required": ["email", "password", "display_name"], "properties": {"email": {"type": "string", "format": "email", "maxLength": 320}, "password": {"type": "string", "minLength": 8, "writeOnly": true}, "display_name": {"type": "string", "minLength": 1, "maxLength": 120}}},
      "LoginRequest": {"type": "object", "required": ["email", "password"], "properties": {"email": {"type": "string", "format": "email"}, "password": {"type": "string", "writeOnly": true}}},
      "TokenResponse": {"type": "object", "required": ["access_token", "token_type", "expires_in"], "properties": {"access_token": {"type": "string"}, "token_type": {"type": "string", "const": "bearer"}, "expires_in": {"type": "integer", "description": "Lifetime in seconds"}}},
      "User": {"type": "object", "required": ["id", "email", "display_name", "status", "created_at", "updated_at"], "properties": {"id": {"type": "string", "format": "uuid"}, "email": {"type": "string", "format": "email"}, "display_name": {"type": "string"}, "status": {"type": "string", "enum": ["active", "disabled"]}, "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}}},
      "Project": {"type": "object", "required": ["id", "owner_id", "name", "created_at", "updated_at"], "properties": {"id": {"type": "string", "format": "uuid"}, "owner_id": {"type": "string", "format": "uuid"}, "name": {"type": "string", "maxLength": 160}, "description": {"type": ["string", "null"]}, "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}}},
      "ProjectCreate": {"type": "object", "required": ["name"], "properties": {"name": {"type": "string", "minLength": 1, "maxLength": 160}, "description": {"type": ["string", "null"], "maxLength": 2000}}},
      "ProjectUpdate": {"type": "object", "minProperties": 1, "properties": {"name": {"type": "string", "minLength": 1, "maxLength": 160}, "description": {"type": ["string", "null"], "maxLength": 2000}}},
      "Membership": {"type": "object", "required": ["project_id", "user_id", "role", "created_at", "updated_at"], "properties": {"project_id": {"type": "string", "format": "uuid"}, "user_id": {"type": "string", "format": "uuid"}, "role": {"$ref": "#/components/schemas/ProjectRole"}, "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}}},
      "MembershipCreate": {"type": "object", "required": ["user_id", "role"], "properties": {"user_id": {"type": "string", "format": "uuid"}, "role": {"$ref": "#/components/schemas/ProjectRole"}}},
      "MembershipUpdate": {"type": "object", "required": ["role"], "properties": {"role": {"$ref": "#/components/schemas/ProjectRole"}}},
      "ProjectRole": {"type": "string", "enum": ["owner", "editor", "viewer"]},
      "TaskStatus": {"type": "string", "enum": ["todo", "in_progress", "done", "archived"]},
      "Task": {"type": "object", "required": ["id", "project_id", "created_by", "title", "status", "priority", "created_at", "updated_at"], "properties": {"id": {"type": "string", "format": "uuid"}, "project_id": {"type": "string", "format": "uuid"}, "created_by": {"type": "string", "format": "uuid"}, "assignee_id": {"type": ["string", "null"], "format": "uuid"}, "title": {"type": "string", "maxLength": 240}, "description": {"type": ["string", "null"], "maxLength": 10000}, "status": {"$ref": "#/components/schemas/TaskStatus"}, "priority": {"type": "integer", "minimum": 0, "maximum": 4}, "due_at": {"type": ["string", "null"], "format": "date-time"}, "created_at": {"type": "string", "format": "date-time"}, "updated_at": {"type": "string", "format": "date-time"}}},
      "TaskCreate": {"type": "object", "required": ["title"], "properties": {"title": {"type": "string", "minLength": 1, "maxLength": 240}, "description": {"type": ["string", "null"], "maxLength": 10000}, "status": {"$ref": "#/components/schemas/TaskStatus"}, "priority": {"type": "integer", "minimum": 0, "maximum": 4}, "assignee_id": {"type": ["string", "null"], "format": "uuid"}, "due_at": {"type": ["string", "null"], "format": "date-time"}}},
      "TaskUpdate": {"type": "object", "minProperties": 1, "properties": {"title": {"type": "string", "minLength": 1, "maxLength": 240}, "description": {"type": ["string", "null"], "maxLength": 10000}, "status": {"$ref": "#/components/schemas/TaskStatus"}, "priority": {"type": "integer", "minimum": 0, "maximum": 4}, "assignee_id": {"type": ["string", "null"], "format": "uuid"}, "due_at": {"type": ["string", "null"], "format": "date-time"}}},
      "WebhookImportRequest": {"type": "object", "required": ["project_id", "tasks"], "properties": {"project_id": {"type": "string", "format": "uuid"}, "actor_id": {"type": ["string", "null"], "format": "uuid", "description": "Optional actor when no Bearer token is supplied; see the threat-model guidance."}, "tasks": {"type": "array", "minItems": 1, "items": {"$ref": "#/components/schemas/TaskCreate"}}}},
      "WebhookImportResponse": {"type": "object", "required": ["event_id", "imported"], "properties": {"event_id": {"type": "string"}, "imported": {"type": "integer", "minimum": 0}}},
      "PageInfo": {"type": "object", "required": ["next_cursor"], "properties": {"next_cursor": {"type": ["string", "null"]}}},
      "ProjectPage": {"type": "object", "required": ["items", "next_cursor"], "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/Project"}}, "next_cursor": {"type": ["string", "null"]}}},
      "MembershipPage": {"type": "object", "required": ["items", "next_cursor"], "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/Membership"}}, "next_cursor": {"type": ["string", "null"]}}},
      "TaskPage": {"type": "object", "required": ["items", "next_cursor"], "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/Task"}}, "next_cursor": {"type": ["string", "null"]}}}
    }
  }
}